Separate RPC API code into multiple files
