+++ aquadoggo/src/errors.rs
@@ -46,7 +46,7 @@ pub enum Error {
 
     /// Custom API errors.
     #[error(transparent)]
-    APIError(#[from] super::rpc::APIError),
+    ApiError(#[from] super::rpc::ApiError),
 }
 
 impl From<Error> for jsonrpc_core::Error {
@@ -74,7 +74,7 @@ impl From<Error> for jsonrpc_core::Error {
             Error::BambooVerification(validation_error) => {
                 handle_validation_error(format!("{}", validation_error))
             }
-            Error::APIError(validation_error) => {
+            Error::ApiError(validation_error) => {
                 handle_validation_error(format!("{}", validation_error))
             }
             _ => {
