+++ aquadoggo/src/rpc/mod.rs
@@ -1,5 +1,9 @@
 mod api;
+mod backend;
+mod request;
+mod response;
 mod server;
 
-pub use api::{APIError, Api, ApiService};
+pub use api::{Api, ApiService};
+pub use backend::ApiError;
 pub use server::RpcServer;
